# contracts/openapi/bff-v1.yaml
openapi: 3.0.3
info:
  title: UnoArena Realtime Gateway / BFF
  version: 1.0.0
  description: |
    Sole external HTTP/SSE surface for UnoArena. Clients submit versioned command
    envelopes, fetch reconnect/projection snapshots, and subscribe to SSE streams.
    Microservices are not publicly reachable. For this implementation the repo-owned
    simple CLI is the sole client/test interface; a graphical UI is deferred and must
    continue to use only this BFF boundary. Rejected commands return `status=rejected`
    with a structured operational/security audit record only (never domain events or
    Game Integrity entries). Unknown or evicted `Last-Event-ID` values return HTTP 409
    `snapshot_required` so clients resync via the snapshot routes before resuming SSE.
servers:
  - url: http://localhost:8080
    description: Local compose BFF (gateway)
...
paths:
...
  /v1/auth/register:
    post:
      tags: [auth]
      summary: Register a player (proxied to Identity)
      operationId: registerPlayer
      parameters:
        - $ref: "#/components/parameters/CorrelationIdHeader"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterRequest"
      responses:
        "200":
          description: Registered
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RegisterResponse"
        "400":
          $ref: "#/components/responses/Error"
        "429":
          $ref: "#/components/responses/Error"
        "502":
          description: Identity upstream failure
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorBody"
  /v1/auth/login:
    post:
      tags: [auth]
      summary: Login and establish a single active session
      operationId: loginPlayer
      parameters:
        - $ref: "#/components/parameters/CorrelationIdHeader"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginRequest"
      responses:
        "200":
          description: Session established
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LoginResponse"
        "401":
          $ref: "#/components/responses/Error"
        "429":
          $ref: "#/components/responses/Error"
        "502":
          description: Identity upstream failure
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorBody"
  /v1/auth/whoami:
    get:
      tags: [auth]
      summary: Resolve the current session principal
      operationId: whoami
      security:
        - SessionBearer: []
      parameters:
        - $ref: "#/components/parameters/CorrelationIdHeader"
        - $ref: "#/components/parameters/SessionTokenHeader"
      responses:
        "200":
          description: Current principal
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WhoamiResponse"
        "401":
          description: Missing or invalid session bearer
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorBody"
        "429":
          $ref: "#/components/responses/Error"
        "502":
          description: Identity upstream failure
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorBody"
  /v1/commands:
    post:
      tags: [commands]
      summary: Submit a versioned command envelope
      description: |
        Compact command entrypoint. `commandId` is the idempotency key.
        `expectedSequenceNumber` is required for mutations of an existing room aggregate.
        Domain rejection is HTTP 200 with `status=rejected` (not HTTP 409).
      operationId: submitCommand
      security:
        - SessionBearer: []
      parameters:
        - $ref: "#/components/parameters/CorrelationIdHeader"
        - $ref: "#/components/parameters/SessionTokenHeader"
        - $ref: "#/components/parameters/CommandIdHeader"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CommandEnvelope"
      responses:
        "200":
          description: Command accepted or rejected (rejection is not a domain event)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CommandResult"
        "400":
          $ref: "#/components/responses/Error"
        "401":
          description: Missing or invalid session bearer
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorBody"
        "429":
          $ref: "#/components/responses/Error"
        "502":
          description: Upstream command dispatch or invalid upstream result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorBody"
  /v1/rooms:
    get:
      tags: [reads]
      summary: Bounded public room list (proxied)
      description: |
        Canonical CLI `room list` / `play --casual` matchmaking read. Transparent Gateway proxy of
        Room Gameplay `GET /internal/v1/rooms/public-list` authenticated with the Gateway↔Room
        scoped service credential. Returns only public room summaries; never private rooms, hands,
        session data, invite material, Game Integrity data, or internal credentials.
        No player bearer is required. Default `status=waiting` (joinable); optional
        `waiting|locked|in_progress`. Opaque Room-owned HMAC cursor; page default 50 / hard max 100.
        Gateway validates query bounds and upstream shape; preserves Room 400/401 classifications
        with Gateway-owned safe error bodies; network/decode/malformed upstream map to 502.
      operationId: listPublicRooms
      parameters:
        - $ref: "#/components/parameters/CorrelationIdHeader"
        - name: status
          in: query
          required: false
          schema:
            type: string
            enum: [waiting, locked, in_progress]
            default: waiting
          description: Room lifecycle filter. Default waiting for joinable casual matchmaking.
        - name: cursor
          in: query
          required: false
          schema:
            type: string
          description: Opaque Room-owned live keyset cursor over stable roomId, bound to the status filter.
        - name: limit
          in: query
          required: false
          schema:
            type: integer
            default: 50
            minimum: 1
            maximum: 100
      responses:
        "200":
          description: Bounded public room summaries
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PublicRoomListPage"
        "400":
          description: Invalid status, limit, or cursor (Room classification preserved)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorBody"
        "401":
          description: Upstream service credential failure (should not reach public clients in healthy wiring)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorBody"
        "429":
          $ref: "#/components/responses/Error"
        "502":
          description: Room upstream failure or malformed public list body
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorBody"
...
  /v1/streams/player:
    get:
      tags: [streams]
      summary: Player SSE feed
      operationId: playerStream
      security:
        - SessionBearer: []
      parameters:
        - $ref: "#/components/parameters/CorrelationIdHeader"
        - $ref: "#/components/parameters/SessionTokenHeader"
        - name: Last-Event-ID
          in: header
          required: false
          schema:
            type: string
          description: Resume marker (event id). Unknown or evicted values yield 409 snapshot_required.
        - name: roomId
          in: query
          required: true
          schema:
            type: string
      responses:
        "200":
          description: |
            text/event-stream of player-safe room facts and control events.
            Wire uses SSE `id`/`event` fields; JSON `data` includes injected `schemaVersion`.
          content:
            text/event-stream:
              schema:
                $ref: "#/components/schemas/SSEEvent"
        "400":
          description: Missing required roomId query
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorBody"
        "401":
          description: Missing or invalid session bearer
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorBody"
        "409":
          $ref: "#/components/responses/SnapshotRequired"
        "429":
          $ref: "#/components/responses/Error"
        "502":
          description: Upstream identity/admission failure while opening the stream
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorBody"
  /v1/streams/spectator:
    get:
      tags: [streams]
      summary: Spectator SSE feed (privacy-filtered)
      description: |
        Admission while room status is `waiting`, `locked`, or `in_progress`. Existing
        streams close on terminal `RoomCompleted` / `RoomCancelled`
        (not on individual `GameCompleted`). When `Authorization` is present, an invalid
        bearer returns HTTP 401 before admission is evaluated.
      operationId: spectatorStream
      security:
        - {}
        - SessionBearer: []
      parameters:
        - $ref: "#/components/parameters/CorrelationIdHeader"
        - $ref: "#/components/parameters/SessionTokenHeader"
        - $ref: "#/components/parameters/RoomInviteHeader"
        - name: Last-Event-ID
          in: header
          required: false
          schema:
            type: string
          description: Resume marker (event id). Unknown or evicted values yield 409 snapshot_required.
        - name: roomId
          in: query
          required: true
          schema:
            type: string
      responses:
        "200":
          description: |
            Spectator-safe SSE stream. Wire uses SSE `id`/`event` fields; JSON `data`
            includes injected `schemaVersion`.
          content:
            text/event-stream:
              schema:
                $ref: "#/components/schemas/SSEEvent"
        "400":
          description: Missing required roomId query
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorBody"
        "401":
          description: Authorization present but session bearer is invalid
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorBody"
        "403":
          description: Admission denied (terminal room or unauthorized)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorBody"
        "409":
          $ref: "#/components/responses/SnapshotRequired"
        "429":
          $ref: "#/components/responses/Error"
        "502":
          description: Spectator admission upstream failure
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorBody"
  /v1/streams/control:
    get:
      tags: [streams]
      summary: Control-plane SSE (session invalidation, reconnect hints)
      operationId: controlStream
      security:
        - SessionBearer: []
      parameters:
        - $ref: "#/components/parameters/CorrelationIdHeader"
        - $ref: "#/components/parameters/SessionTokenHeader"
        - name: Last-Event-ID
          in: header
          required: false
          schema:
            type: string
          description: Resume marker (event id). Unknown or evicted values yield 409 snapshot_required.
      responses:
        "200":
          description: |
            Control SSE stream. Wire uses SSE `id`/`event` fields; JSON `data`
            includes injected `schemaVersion`.
          content:
            text/event-stream:
              schema:
                $ref: "#/components/schemas/SSEEvent"
        "401":
          description: Missing or invalid session bearer
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorBody"
        "409":
          $ref: "#/components/responses/SnapshotRequired"
        "429":
          $ref: "#/components/responses/Error"
        "502":
          description: Upstream identity failure while opening the control stream
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorBody"
  /v1/rankings/leaderboards:
    get:
      tags: [reads]
      summary: Public leaderboard read model (proxied)
      operationId: getLeaderboard
      parameters:
        - $ref: "#/components/parameters/CorrelationIdHeader"
        - name: boardType
          in: query
          required: true
          schema:
            type: string
            enum: [casual_elo, tournament_placement]
        - name: cursor
          in: query
          required: false
          schema:
            type: string
          description: Opaque Ranking-owned live keyset cursor over the last (rating, playerId) boundary. A projection version change does not invalidate it.
        - name: limit
          in: query
          required: false
          schema:
            type: integer
            default: 100
            minimum: 1
            maximum: 500
      responses:
        "200":
          description: Leaderboard snapshot
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LeaderboardPage"
        "429":
          $ref: "#/components/responses/Error"
        "502":
          description: Ranking upstream failure
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorBody"
...
  /v1/analytics/public:
    get:
      tags: [reads]
      summary: Public analytics read model (proxied)
      operationId: getPublicAnalytics
      parameters:
        - $ref: "#/components/parameters/CorrelationIdHeader"
      responses:
        "200":
          description: Derived public analytics
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
        "429":
          $ref: "#/components/responses/Error"
        "502":
          description: Analytics upstream failure
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorBody"
components:
  securitySchemes:
    SessionBearer:
      type: http
      scheme: bearer
      bearerFormat: opaque-session-token
      description: |
        Identity-issued session token via BFF login. Required for player/command/control
        routes. Optional on spectator routes; when present, invalid tokens yield 401.
  parameters:
    CorrelationIdHeader:
      name: X-Correlation-Id
      in: header
      required: false
      schema:
        type: string
    CommandIdHeader:
      name: X-Command-Id
      in: header
      required: false
      schema:
        type: string
    SessionTokenHeader:
      name: Authorization
      in: header
      required: false
      schema:
        type: string
      description: Bearer session token issued by Identity via BFF login
    RoomInviteHeader:
      name: X-Room-Invite
      in: header
      required: false
      schema:
        type: string
      description: Invite capability token for private-room spectator admission
  responses:
    Error:
      description: Structured error
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorBody"
    SnapshotRequired:
      description: |
        Last-Event-ID is unknown or evicted from the resume bound. Client must fetch
        the matching snapshot route and resume SSE without claiming false continuity.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/SnapshotRequiredError"
          example:
            code: snapshot_required
            message: Last-Event-ID unknown or evicted; resync snapshot required
            correlationId: corr_01HX...
  schemas:
    PublicRoomListPage:
      type: object
      additionalProperties: false
      required: [schemaVersion, rooms]
      properties:
        schemaVersion:
          type: integer
          enum: [1]
        rooms:
          type: array
          maxItems: 100
          items:
            $ref: "#/components/schemas/PublicRoomSummary"
        nextCursor:
          type: string
          description: Present only when another page exists (one-row lookahead).
...
    RegisterRequest:
      type: object
      required: [username, password]
      properties:
        username:
          type: string
        password:
          type: string
          format: password
    RegisterResponse:
      type: object
      required: [status, username]
      properties:
        status:
          type: string
        username:
          type: string
    LoginRequest:
      type: object
      required: [username, password]
      properties:
        username:
          type: string
        password:
          type: string
          format: password
    LoginResponse:
      type: object
      required: [sessionId, playerId, token]
      properties:
        sessionId:
          type: string
        playerId:
          type: string
        token:
          type: string
    WhoamiResponse:
      type: object
      required: [playerId, sessionId, username]
      properties:
        playerId:
          type: string
        sessionId:
          type: string
        username:
          type: string
    CommandEnvelope:
      description: |
        Discriminated command envelope. `type` selects exactly one catalog variant;
        each variant constrains `payload` and whether `expectedSequenceNumber` is
        required (existing-room mutations) or prohibited (CreateRoom and tournament
        commands). Gateway accepts only these envelope fields.
      oneOf:
        - $ref: "#/components/schemas/CreateRoomCommand"
        - $ref: "#/components/schemas/JoinRoomCommand"
        - $ref: "#/components/schemas/LeaveRoomCommand"
        - $ref: "#/components/schemas/LockRoomCommand"
        - $ref: "#/components/schemas/StartMatchCommand"
        - $ref: "#/components/schemas/CancelRoomCommand"
        - $ref: "#/components/schemas/PlayCardCommand"
        - $ref: "#/components/schemas/DrawCardCommand"
        - $ref: "#/components/schemas/ChooseColorCommand"
        - $ref: "#/components/schemas/CallUnoCommand"
        - $ref: "#/components/schemas/ReportMissingUnoCommand"
        - $ref: "#/components/schemas/ReconnectToRoomCommand"
        - $ref: "#/components/schemas/CreateTournamentCommand"
        - $ref: "#/components/schemas/RegisterPlayerCommand"
        - $ref: "#/components/schemas/CloseRegistrationCommand"
      discriminator:
        propertyName: type
        mapping:
          CreateRoom: "#/components/schemas/CreateRoomCommand"
          JoinRoom: "#/components/schemas/JoinRoomCommand"
          LeaveRoom: "#/components/schemas/LeaveRoomCommand"
          LockRoom: "#/components/schemas/LockRoomCommand"
          StartMatch: "#/components/schemas/StartMatchCommand"
          CancelRoom: "#/components/schemas/CancelRoomCommand"
          PlayCard: "#/components/schemas/PlayCardCommand"
          DrawCard: "#/components/schemas/DrawCardCommand"
          ChooseColor: "#/components/schemas/ChooseColorCommand"
          CallUno: "#/components/schemas/CallUnoCommand"
          ReportMissingUno: "#/components/schemas/ReportMissingUnoCommand"
          ReconnectToRoom: "#/components/schemas/ReconnectToRoomCommand"
          CreateTournament: "#/components/schemas/CreateTournamentCommand"
          RegisterPlayer: "#/components/schemas/RegisterPlayerCommand"
          CloseRegistration: "#/components/schemas/CloseRegistrationCommand"
...
    CommandResult:
      type: object
      required: [commandId, type, status, schemaVersion]
      properties:
        commandId:
          type: string
        type:
          type: string
        status:
          type: string
          enum: [accepted, rejected]
        schemaVersion:
          type: integer
          enum: [1]
        reason:
          type: string
          description: Present when status=rejected; pairs with operational audit record
        sequenceNumber:
          type: integer
          format: int64
        payload:
          type: object
          additionalProperties: true
...
    ErrorBody:
      type: object
      required: [code, message]
      properties:
        code:
          type: string
        message:
          type: string
        correlationId:
          type: string
        commandId:
          type: string
    SnapshotRequiredError:
      allOf:
        - $ref: "#/components/schemas/ErrorBody"
        - type: object
          required: [code, message]
          properties:
            code:
              type: string
              enum: [snapshot_required]
            message:
              type: string
              example: Last-Event-ID unknown or evicted; resync snapshot required
...